Improve the organization of the dtype mapping code
