+++ merlin/dtype/mappings/python.py
@@ -1,0 +1,29 @@
+#
+# Copyright (c) 2022, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import numpy as np
+
+from merlin.dtype import dtypes
+from merlin.dtype.registry import _dtype_registry
+
+python_dtypes = {
+    dtypes.boolean: bool,
+    dtypes.int64: int,
+    dtypes.float64: float,
+    dtypes.string: str,
+}
+_dtype_registry.register("python", python_dtypes)
+
