+++ merlin/dtype/registry.py
@@ -15,50 +15,7 @@
 #
 from typing import Dict, Union
 
-
-class DTypeMapping:
-    def __init__(self, mapping, base_class=None):
-        self.from_merlin_ = mapping
-        self.to_merlin_ = {}
-        self.base_class = base_class
-
-        for key, values in mapping.items():
-            if not isinstance(values, list):
-                values = [values]
-            for value in values:
-                self.to_merlin_[value] = key
-
-    def matches_external(self, external_dtype):
-        return self._matches(external_dtype, self.to_merlin_, self.base_class)
-
-    def matches_merlin(self, merlin_dtype):
-        return self._matches(merlin_dtype, self.from_merlin_)
-
-    def to_merlin(self, external_dtype, shape=None):
-        merlin_dtype = self.to_merlin_[external_dtype]
-        return merlin_dtype
-
-    def from_merlin(self, merlin_dtype):
-        # Always translate to the first external dtype in the list
-        return self.from_merlin_[merlin_dtype][0]
-
-    def _matches(self, dtype, mapping, base_class=None):
-        # If the mapping requires that the dtype is a subclass
-        # of a particular base class and it isn't, then we
-        # can immediately fail to match and exit.
-        if base_class and not isinstance(dtype, base_class):
-            return False
-
-        # Some external dtype objects are not hashable, so they
-        # can't be used as dictionary keys. In that case, match
-        # against the dtype class instead.
-        hashable_dtype = dtype
-        try:
-            hash(dtype)
-        except TypeError:
-            hashable_dtype = type(dtype)
-
-        return hashable_dtype in mapping.keys()
+from merlin.dtype.mapping import DTypeMapping
 
 
 class DTypeMappingRegistry:
